Move conf.py to the wagtailtrans root, utils is to generic
