+++ src/wagtailtrans/signals.py
@@ -3,10 +3,10 @@ from django.db.models.signals import m2m_changed, post_save, pre_delete
 from wagtail.wagtailadmin.signals import init_new_page
 from wagtail.wagtailcore.models import Site, get_page_models
 
+from wagtailtrans.conf import get_wagtailtrans_setting
 from wagtailtrans.models import Language, SiteLanguages, TranslatablePage
 from wagtailtrans.permissions import (
     create_group_permissions, get_or_create_language_group)
-from wagtailtrans.utils.conf import get_wagtailtrans_setting
 
 
 def synchronize_trees(sender, instance, **kwargs):
